+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: alpayet <alpayet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 11:26:16 by alpayet           #+#    #+#             */
-/*   Updated: 2024/11/09 00:16:17 by alpayet          ###   ########.fr       */
+/*   Updated: 2024/11/22 01:28:34 by alpayet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
